Fix minibatch advi type casting

+++ pymc3_models/models/HierarchicalLogisticRegression.py
@@ -29,7 +29,7 @@ class HierarchicalLogisticRegression(BayesianModel):
         """
         model_input = theano.shared(np.zeros([self.num_training_samples, self.num_pred]))
 
-        model_output = theano.shared(np.zeros(self.num_training_samples))
+        model_output = theano.shared(np.zeros(self.num_training_samples, dtype='int'))
 
         model_cats = theano.shared(np.zeros(self.num_training_samples, dtype='int'))
 
@@ -128,7 +128,11 @@ class HierarchicalLogisticRegression(BayesianModel):
         if self.cached_model is None:
             self.cached_model = self.create_model()
 
-        self._set_shared_vars({'model_input': X, 'model_output': np.zeros(num_samples), 'model_cats': cats})
+        self._set_shared_vars({
+            'model_input': X,
+            'model_output': np.zeros(num_samples, dtype='int'),
+            'model_cats': cats
+        })
 
         ppc = pm.sample_ppc(self.trace, model=self.cached_model, samples=2000)
